Refactor : 전체적인 구조 변환
