+++ mission-racing/src/main/java/com/racing/common/domain/Cars.java
@@ -56,8 +56,4 @@ public class Cars {
                 .map(Car::getMoveCount)
                 .collect(Collectors.toList());
     }
-
-    public List<Car> getCars() {
-        return cars;
-    }
 }
