+++ mission-racing/src/main/java/com/racing/console/controller/RacingController.java
@@ -2,7 +2,7 @@ package com.racing.console.controller;
 
 import com.racing.common.domain.Cars;
 import com.racing.common.domain.NumberGenerator;
-import com.racing.common.view.dto.ResultResponse;
+import com.racing.common.dto.ResultResponse;
 import com.racing.console.view.InputView;
 import com.racing.console.view.OutputView;
